Fixed issue with git misstracking file name
