+++ src/main/java/frc/robot/commands/AlignmentRightPeg.java
@@ -4,22 +4,23 @@ import com.pathplanner.lib.auto.AutoBuilder;
 import com.pathplanner.lib.path.PathConstraints;
 
 import edu.wpi.first.math.geometry.Pose2d;
+import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.Constants.POSES;
 import frc.robot.LimelightHelpers;
 import frc.robot.subsystems.Swerve;
 /* You should consider using the more terse Command factories API instead https://docs.wpilib.org/en/stable/docs/software/commandbased/organizing-command-based.html#defining-commands */
-public class AlignmentLeft extends Command {
+public class AlignmentRightPeg extends Command {
   Swerve s_Swerve;
   boolean isrightscore;
+
 
         boolean isFinished = false;
 
 
     private Pose2d Targetpose = null;
-    public static int lastpeg = 0;
 
-  public AlignmentLeft(Boolean isrightscore, Swerve s_Swerve) {
+  public AlignmentRightPeg(Boolean isrightscore, Swerve s_Swerve) {
     this.s_Swerve = s_Swerve;
     this.isrightscore = isrightscore;
   }
@@ -28,7 +29,7 @@ public class AlignmentLeft extends Command {
   @Override
  public void initialize() {
         
-            System.out.println("leftBranchPathfinding method called.");
+            System.out.println("rightBranchPathfinding method called.");
         
             double aprilTagID = LimelightHelpers.getFiducialID("limelight");
             // Blue paths
@@ -36,87 +37,75 @@ public class AlignmentLeft extends Command {
             switch ((int) aprilTagID) {
             case 17:
                  {
-                    Targetpose = POSES.REEF_C;
-                    lastpeg = 2;
+                    Targetpose = POSES.REEF_D;
                 }
                 break;
 
             case 18 :
             {
-                Targetpose = POSES.REEF_A;
-                lastpeg = 0;
+                Targetpose = POSES.REEF_B;
             }
                 break;
             
 
             case 19:
             {
-                 Targetpose = POSES.REEF_K;
-                 lastpeg = 10;
+                 Targetpose = POSES.REEF_L;
             }
                 break;
 
             case 20:
             {
-                 Targetpose = POSES.REEF_I;
-                 lastpeg = 8;
+                 Targetpose = POSES.REEF_J;
             }
                 break;
 
             case 21:
             {
-                 Targetpose = POSES.REEF_G;
-                 lastpeg = 6;
+                 Targetpose = POSES.REEF_H;
             }
                 break;
 
             case 22:
             {
-                Targetpose = POSES.REEF_E;
-                lastpeg = 4;
+                 Targetpose = POSES.REEF_F;
             }
                 break;
 
             // Red Paths
             case 6:
             {
-                 Targetpose = POSES.REEF_K;
-                 lastpeg = 10;
+                 Targetpose = POSES.REEF_L;
             }
                 break;
 
             case 7:
             {
-                 Targetpose = POSES.REEF_A;
-                 lastpeg = 0;
+                 Targetpose = POSES.REEF_B;
             }
                 break;
 
             case 8:
             {
-                 Targetpose = POSES.REEF_C;
-                 lastpeg = 2;
+                 Targetpose = POSES.REEF_D;
             }
                 break;
 
             case 9:
             {
-                 Targetpose = POSES.REEF_E;
-                 lastpeg = 4;
+                 Targetpose = POSES.REEF_F;
             }
                 break;
 
             case 10:
             {
-                 Targetpose = POSES.REEF_G;
-                 lastpeg = 6;
+                 Targetpose = POSES.REEF_H;
             }
                 break;
 
             case 11:
                  {
-                 Targetpose = POSES.REEF_I;
-                 lastpeg = 8;
+                 Targetpose = POSES.REEF_J;
                
                 }
                 break;
@@ -133,22 +122,30 @@ public class AlignmentLeft extends Command {
           
           @Override
           public void execute() {
-            addRequirements(s_Swerve);
         
-            
+            s_Swerve.drive(new Translation2d(0,0), 0, false,false);
             PathConstraints constraints= new PathConstraints(
+            1,
         
-                    1,
-        
-                    4,
-        
-                    9.42478,
-        
-                    12.5664
+            4,
+
+            9.42478,
+
+            12.5664
         
             );
 
+    //         = new PathConstraints(
         
+    //         5,
+
+    //         3,
+
+    //         4,
+
+    //         3
+
+    // );
 
     if (Targetpose != null) {
 
@@ -156,7 +153,7 @@ public class AlignmentLeft extends Command {
         Command followLeftPath = AutoBuilder.pathfindToPose(
             Targetpose,
             constraints,
-            0.00
+            0.01
     );
             followLeftPath.schedule();
             
@@ -179,6 +176,7 @@ public class AlignmentLeft extends Command {
 
 
     cancel();
+
   }
     
     
@@ -195,9 +193,5 @@ public class AlignmentLeft extends Command {
         return false;
     }
 
-}
-
-public static class lastpegsave {
-    public static int lastpegsaved = lastpeg;
 }
 }
